mpvacious: allow custom mpv flags on extraction
Add two new configuration fields (mpv_extract_{audio,snapshot}_flags),
which are appended to the mpv argument list when an audio clip or
snapshot (respectively) are being generated. This allows you to apply
fairly complicated filters to the extracted data (the most obvious
usecase being to apply loudness normalisation filters to make the volume
of all cards the same).

Note that because these filters are applied in one-shot, filters like
silenceremove will cause the start and end of the clip to be incorrect.
The next patch will add a generic mechanism for post-processing clips.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>